fix(api): wire token registry into node startup and harden reconciliation
Addresses the second PR review round. The first finding made the whole
feature inert, and the review was right to lead with it.

Wire the config into the production composition. internal/node's
buildDingoConfig maps internal config onto dingo.With* options field by
field, and there was no WithTokenRegistryConfig call — so an operator's
tokenRegistry block parsed, validated, and was then dropped, leaving
Enabled false and the sync never starting. My earlier test used
NewConfigFromInternal, a different path, which is exactly why it passed
while production was broken. The new test exercises buildDingoConfig.

Truncate the snapshot stamp to a whole second. MySQL's `datetime`
carries no fractional seconds, so a sub-second stamp was stored rounded
while the prune compared against the original — every row the snapshot
had just written looked older than the cutoff and would have been
deleted. Truncating at the source makes written value and cutoff
identical on all three backends.

Never reconcile against an archive with no mapping files. That is what a
layout change, a truncated artifact, or a mirror serving the wrong
repository looks like, and pruning against it would empty the table and
record the ETag so it stuck. The discriminator is the mapping-file count
rather than the entry count: an archive that does carry mappings which
all turn out unusable is a real empty registry and must still reconcile.

Defer reconciliation when any mapping was skipped. A skipped mapping
does not re-stamp its row, so a mapping that was valid yesterday and is
malformed or oversized today would have had its still-good metadata
deleted.

Namespace the stored ETag by source URL. A validator means nothing to a
source that did not issue it, so repointing sourceUrl or switching
network could otherwise accept a 304 for the wrong registry.

Reject an explicit null decimals. encoding/json treats null into an int
as a no-op returning no error, so `"value": null` was published as a
declared 0 — which a wallet would use to render balances unscaled.

Cap decimals at 19, not the 30 I invented: CIP-26's schema declares
{"minimum": 0, "maximum": 19}. Boundary tests pin both ends.

Also unifies the two sibling store methods on `int` and removes two
dead test helpers left behind by modernize's new(expr) rewrite.

Declined: rejecting properties whose envelope omits `sequenceNumber`.
Nothing here consumes that field, so dropping otherwise-valid metadata
over it would trade real data for no safety gain.

+++ database/plugin/metadata/sqlite/shared_sqlstore_tokenregistry_test.go
@@ -212,7 +212,7 @@ type tokenRegistryPruneStore interface {
 		context.Context,
 		time.Time,
 		types.Txn,
-	) (int64, error)
+	) (int, error)
 }
 
 // TestSharedSQLStoreTokenRegistryPrune covers the reconciliation half of a
@@ -252,7 +252,7 @@ func TestSharedSQLStoreTokenRegistryPrune(t *testing.T) {
 	pruned, err := store.PruneTokenRegistryEntriesBefore(ctx, secondSync, nil)
 
 	require.NoError(t, err)
-	require.Equal(t, int64(1), pruned)
+	require.Equal(t, 1, pruned)
 	survivor, err := store.GetTokenRegistryEntry(testSubjectNut, nil)
 	require.NoError(t, err)
 	require.NotNil(t, survivor)
